refactor(lsp): rename legacy sumneko_lua references to lua_ls
The Lua language server was renamed from sumneko_lua to lua_ls a while
ago. The setup code already used lua_ls, but the settings file and
local variable still carried the old name. This aligns naming across
the config and fixes a typo in the commented basedpyright require.
